pyproject.toml
[tool.poetry]
name = "rfid_macros"
version = "0.1.0"
description = ""
authors = ["Joey Stevens <joey.stevens00@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
evdev = "^1.3.0"
pyyaml = "^5.3.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
